requirements.txt
Django==1.7.3
django-toolbelt==0.0.1
wsgiref==0.1.2
python-ldap==2.4